キーボードが点滅する。ImageかPanelが必要だが
